fix: 금액 단위·정밀도 계약을 원 단위 정수(long/BIGINT)로 통일 (NB2)
입찰가→PG 응답 검증→결제 저장 전 구간의 금액 단위·정밀도 계약을 명문화한다.
기존은 PG 응답을 intValue()로 절삭, 입찰가를 (int)(price*100)로 캐스팅, 결제를
다시 Double로 저장해 조용한 절삭·범위 초과·부동소수점 경로가 있었다.

- 도메인·DTO 금액 필드를 Double/double → Long/long(원 단위 정수)으로 전환
  (Music.startingPrice/currentHighestBid, Bid.price, Payment.price, 관련 DTO,
   가격 필터 파라미터, QueryDSL NumberExpression<Long>, comparingLong)
- KrwAmounts: 원(KRW) 최소 단위 1원·반올림/절삭 금지 정책 명문화.
  PG 금액(BigDecimal)을 longValueExact로 무손실 변환(소수·범위 초과는 예외)
- PaymentGateway 경계 도입(IamportPaymentGateway 실구현) — PG 호출을 추상화해
  네트워크 없이 대역으로 결정적 검증 가능. 결제 검증에서 *100 배율 제거(원 단위 직접 비교)
- 마이그레이션: scripts/migration/nb2-01-amount-to-bigint.sql
  (preflight로 소수·범위 초과 행 탐지→중단, 백업, DOUBLE→BIGINT, 전후 보존 검증)
- 테스트 13종 추가: KrwAmounts 절삭/오버플로우 재현+무손실 계약, PaymentService
  금액 일치/불일치/소유권(Mockito), 입찰→낙찰→PG 대역→결제 저장→정산 통합(H2)
- 기존 계약/특성화 테스트를 새 타입에 맞게 갱신

프론트 수정·브라우저 E2E·실제 PG 결제는 범위 밖(백엔드 계약 기준). 기존 프론트의
"만원"·*100 표시는 신규 API 단위와 비호환이며 프론트 호환은 별도 후속 작업.

+++ src/main/java/com/notenest/controller/MusicController.java
@@ -109,8 +109,8 @@ public class MusicController {
     public ResponseEntity<Page<MusicSummaryDTO>> getAllMusic(
             @RequestParam(required = false) String majorGenre,
             @RequestParam(required = false) String hashtag,
-            @RequestParam(required = false) Double minPrice,
-            @RequestParam(required = false) Double maxPrice,
+            @RequestParam(required = false) Long minPrice,
+            @RequestParam(required = false) Long maxPrice,
             @RequestParam(defaultValue = "0") int page,
             @RequestParam(defaultValue = "10") int size,
             @RequestParam(defaultValue = "latest") String sortBy,
