+++ src/main/java/com/notenest/dto/CreateMusicDTO.java
@@ -11,7 +11,7 @@ public class CreateMusicDTO {
     private String title;
     private String subtitle;
     private String majorGenre;
-    private Double startingPrice;
+    private Long startingPrice;
     private Integer musicPeriod;
     private String details;
     private String hashtag;
@@ -19,7 +19,7 @@ public class CreateMusicDTO {
     private byte[] audio;
     private UUID userUuid; // 음악을 등록한 사용자의 UUID
     private String nickName; // 음악을 등록한 사용자의 닉네임
-    private Double currentHighestBid;
+    private Long currentHighestBid;
     private LocalDateTime auctionEndTime;
     private Boolean showAllBids; // 입찰 내역을 모두 보여줄지 여부
     private Boolean popularComposer;
